Task 5-Personal Finance Calculator

+++ python_introduction/finance_calculator.py
@@ -3,7 +3,7 @@ monthly_income = int(input('Enter your monthly income: '))
 monthly_expense = int(input('Enter your total monthly expenses: '))
 
 monthly_saving = int(monthly_income - monthly_expense)
-projected_savings = int(monthly_saving * 12 + (monthly_saving * 12 * 0.05))
+projected_savings = int(monthly_saving * 12 + int(monthly_saving * 12 * 0.05))
 
 print(f'Enter your monthly income: {monthly_income}')
 print(f'Enter your total monthly expenses: {monthly_expense}')
